split out dropdown options into a dedicated element
